+++ Backend/src/main/java/com/pastelarte/pastelarte_api/dto/PedidoRequestDTO.java
@@ -1,7 +1,9 @@
+// PedidoRequestDTO.java
 package com.pastelarte.pastelarte_api.dto;
 
 import lombok.*;
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -12,7 +14,7 @@ public class PedidoRequestDTO {
     private Integer idCliente;
     private LocalDate fecha;
     private String estado;
-    private Double total;
+    private BigDecimal total;
     private Integer idPago;
     private String comprobante;
     private String numeroNequi;
