+++ Backend/src/main/java/com/pastelarte/pastelarte_api/dto/PersonalizacionResponseDTO.java
@@ -1,6 +1,8 @@
+// PersonalizacionResponseDTO.java
 package com.pastelarte.pastelarte_api.dto;
 
 import lombok.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,5 +15,5 @@ public class PersonalizacionResponseDTO {
     private String sabor;
     private String decoraciones;
     private String descripcion;
-    private Float costoExtra;
+    private BigDecimal costoExtra;
 }
