+++ Backend/src/main/java/com/pastelarte/pastelarte_api/dto/ProductoRequestDTO.java
@@ -1,6 +1,8 @@
+// ProductoRequestDTO.java
 package com.pastelarte.pastelarte_api.dto;
 
 import lombok.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -10,7 +12,7 @@ public class ProductoRequestDTO {
 
     private String nombre;
     private String descripcion;
-    private Float precio;
+    private BigDecimal precio;
     private String categoria;
     private String imagen;
 }
