+++ Backend/src/main/java/com/pastelarte/pastelarte_api/dto/ProductoResponseDTO.java
@@ -1,6 +1,8 @@
+// ProductoResponseDTO.java
 package com.pastelarte.pastelarte_api.dto;
 
 import lombok.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -11,7 +13,7 @@ public class ProductoResponseDTO {
     private Integer idProducto;
     private String nombre;
     private String descripcion;
-    private Float precio;
+    private BigDecimal precio;
     private String categoria;
     private String imagen;
 }
